`precision=double` fixes Added the ability to disable world coordinates for shaders. Added the ability to fix the position of instances. `ScopedPairIdConfig` is now a structure. Added a class `ViewportToDebugContainerItem` for storing containers of the world and viewports. Replaced `typedef` with `using`. Fixed warnings about converting `double` to `float`.

+++ src/3d/render_instances.cpp
@@ -512,7 +512,7 @@ GeometryType GeometryPool::_scoped_config_get_geometry_type(const std::shared_pt
 Color GeometryPool::_scoped_config_to_custom(const std::shared_ptr<DebugDraw3DScopeConfig::Data> &p_cfg) {
 	// ZoneScoped;
 	if (_scoped_config_get_geometry_type(p_cfg) == GeometryType::Volumetric)
-		return Color(p_cfg->thickness, p_cfg->center_brightness, 0, 0);
+		return Color((float)p_cfg->thickness, (float)p_cfg->center_brightness, (float)0, (float)0);
 
 	return Color();
 }
